rename declaration file
