+++ src/routes/safe/components/Apps/hooks/useIframeMessageHandler.ts
@@ -19,7 +19,7 @@ import {
   safeParamAddressFromStateSelector,
 } from 'src/logic/safe/store/selectors'
 import { TransactionParams } from '../components/AppFrame'
-import { SafeApp } from 'src/routes/safe/components/Apps/types.d'
+import { SafeApp } from 'src/routes/safe/components/Apps/types'
 
 type InterfaceMessageProps<T extends InterfaceMessageIds> = {
   messageId: T
